Integrate Code Together
