+++ application/bus_me/lib/views/login.dart
@@ -1,108 +1,80 @@
 
 import 'package:bus_me/observable.dart';
+import 'package:bus_me/views/admin_view.dart';
+import 'package:bus_me/views/map_view.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
-class LoginPage extends StatefulWidget
-{
-  final Observable loginObservable = BaseObservable();
-
-  LoginPage({Key? key}):
-        super(key: key);
-
+class LoginScreen extends StatefulWidget {
   @override
-  State<StatefulWidget> createState() => LoginPageState();
+  _LoginScreenState createState() => _LoginScreenState();
 }
 
-class LoginPageState extends State<LoginPage>
-{
-
-  TextEditingController usernameController = TextEditingController();
-  TextEditingController passwordController = TextEditingController();
+class _LoginScreenState extends State<LoginScreen> {
+  final _usernameController = TextEditingController();
+  final _passwordController = TextEditingController();
 
   @override
-  Widget build(BuildContext context)
-  {
-    return
-      Padding(
-        padding: const EdgeInsets.all(20),
-        child: ListView(
-          //all children of login menu
-          children: [
-
-            //title
-            Container(
-              padding: const EdgeInsets.all(10),
-              alignment: Alignment.center,
-              child: const Text(
-                'BusME',
-                style: TextStyle(
-                  fontSize: 48
-                ),
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text('BusMe'),
+      ),
+      body: Padding(
+        padding: const EdgeInsets.all(16.0),
+        child: Column(
+          mainAxisAlignment: MainAxisAlignment.center,
+          children: <Widget>[
+            TextField(
+              controller: _usernameController,
+              decoration: InputDecoration(
+                labelText: 'Username',
+                border: OutlineInputBorder(),
               ),
             ),
-
-            //login text
-            Container(
-              padding: const EdgeInsets.all(10),
-              alignment: Alignment.center,
-              child: const Text(
-                'Login',
-                style: TextStyle(
-                    fontSize: 18
-                ),
+            SizedBox(height: 16.0),
+            TextField(
+              controller: _passwordController,
+              obscureText: true,
+              decoration: InputDecoration(
+                labelText: 'Password',
+                border: OutlineInputBorder(),
               ),
             ),
+            SizedBox(height: 24.0),
+            ElevatedButton(
+              onPressed: () {
+                String username = _usernameController.text;
+                String password = _passwordController.text;
 
-            //login boxes
-            //username field
-            Container(
-              padding: const EdgeInsets.all(10),
-              alignment: Alignment.center,
-              child: TextField(
-                controller: usernameController,
-                decoration: const InputDecoration(
-                  border: OutlineInputBorder(),
-                  labelText: "Username"
-                ),
-              )
-            ),
-
-            Container(
-                padding: const EdgeInsets.all(10),
-                alignment: Alignment.center,
-                child: TextField(
-                  obscureText: true,
-                  controller: passwordController,
-                  decoration: const InputDecoration(
-                      border: OutlineInputBorder(),
-                      labelText: "Password"
-                  ),
-                )
-            ),
-            //enty button
-            Container(
-
-                padding: const EdgeInsets.fromLTRB(10, 0, 10 , 0),
-                alignment: Alignment.center,
-                child: ElevatedButton(
-                  style: ElevatedButton.styleFrom(minimumSize:  const Size.fromHeight(40)),
-                  onPressed:  () async {
-                    await widget.loginObservable.notifyObservers(
-                        ObsSignal("login", {"username": usernameController.text, "password": passwordController.text})
-                    );
-                  },
-                  child: const Text(
-                    'login',
-                    style: TextStyle(
-                      fontSize: 18
+                if (username == "admin") {
+                  Navigator.push(
+                    context,
+                    MaterialPageRoute(
+                      builder: (context) => AdminView(),
                     ),
-                  ),
-                )
+                  );
+                } else {
+                  Navigator.push(
+                    context,
+                    MaterialPageRoute(
+                      builder: (context) => MapView(),
+                    ),
+                  );
+                }
+              },
+              child: Text('Login'),
             ),
-
           ],
-        )
-      );
+        ),
+      ),
+    );
+  }
+
+  @override
+  void dispose() {
+    _usernameController.dispose();
+    _passwordController.dispose();
+    super.dispose();
   }
 }
