+++ application/bus_me/lib/views/map_view.dart
@@ -1,3 +1,4 @@
+
 import 'package:flutter/material.dart';
 import 'package:flutter_map/flutter_map.dart';
 import 'package:latlong2/latlong.dart';
@@ -64,3 +65,6 @@ class MapView extends StatelessWidget {
         ));
   }
 }
+
+
+
